refactor: 권한별 uri 분기
