+++ src/README.md
@@ -75,26 +75,50 @@
 
 ## API 명세서
 
+### 페이지
+
+#### 관리자
+
+| HTTP Method | URI                  | Description  |
+|-------------|----------------------|--------------|
+| GET         | `/admin`             | 관리자 메인 페이지   | 
+| GET         | `/admin/reservation` | 예약 관리 페이지    |
+| GET         | `/admin/time`        | 예약 시간 관리 페이지 |
+| GET         | `/admin/theme`       | 테마 관리 페이지    |
+
+#### 그 외
+
+| HTTP Method | URI                 | Description  |
+|-------------|---------------------|--------------|
+| GET         | `/`                 | 사용자 메인 페이지   |
+| GET         | `/reservation`      | 예약 페이지       |
+| GET         | `/login`            | 로그인 페이지      |
+| GET         | `/reservation-mine` | 예약 목록 조회 페이지 |
+
+### API
+
+#### 관리자
+
+| HTTP Method | URI                          | Description    |
+|-------------|------------------------------|----------------|
+| GET         | `/admin/reservations`        | 예약 목록 조회       |
+| POST        | `/admin/reservations`        | 관리자 예약 추가      |
+| DELETE      | `/admin/reservations/{id}`   | 예약 삭제          |
+| GET         | `/admin/reservations/filter` | 예약 목록 필터링      |
+| POST        | `/admin/times`               | 예약 시간 추가       |
+| DELETE      | `/admin/times/{id}`          | 예약 시간 삭제       |
+| POST        | `/admin/themes`              | 테마 추가          |
+| DELETE      | `/admin/themes/{id}`         | 테마 삭제          |
+| GET         | `/admin/members`             | 간소화된 사용자 목록 조회 |
+
+#### 그 외
+
 | HTTP Method | URI                     | Description       |
 |-------------|-------------------------|-------------------|
-| GET         | `/admin`                | 관리자 메인 페이지        | 
-| GET         | `/admin/reservation`    | 예약 관리 페이지         |
-| GET         | `/admin/time`           | 예약 시간 관리 페이지      |
-| GET         | `/admin/theme`          | 테마 관리 페이지         |
-| GET         | `/`                     | 사용자 메인 페이지        |
-| GET         | `/reservation`          | 예약 페이지            |
-| GET         | `/login`                | 로그인 페이지           |
-| GET         | `/reservation-mine`     | 예약 목록 조회 페이지      |
-| GET         | `/reservations`         | 예약 목록 조회          |
+| GET         | `/reservations`          | 사용자가 예약한 예약 목록 조회 |
 | POST        | `/reservations`         | 예약 추가             |
-| GET         | `/reservations/filter`  | 예약 목록 필터링         |
-| DELETE      | `/reservations/{id}`    | 예약 삭제             |
 | GET         | `/times`                | 예약 시간 목록 조회       |
-| POST        | `/times`                | 예약 시간 추가          |
-| DELETE      | `/times/{id}`           | 예약 시간 삭제          |
 | GET         | `/themes`               | 테마 목록 조회          |
-| POST        | `/themes`               | 테마 추가             |
-| DELETE      | `/themes/{id}`          | 테마 삭제             | 
 | POST        | `/login`                | 로그인 요청            |
 | GET         | `/login/check`          | 로그인 정보 조회         |
-| GET         | `/members/reservations` | 사용자가 예약한 예약 목록 조회 |
+| POST        | `/logout`               | 로그아웃 요청           |
