--- src/main/java/roomescape/controller/api/AdminController.java
@@ -1,32 +1,0 @@
-package roomescape.controller.api;
-
-import java.net.URI;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import roomescape.dto.request.AdminReservationRequest;
-import roomescape.dto.response.ReservationResponse;
-import roomescape.service.ReservationService;
-
-@RestController
-@RequestMapping("/admin")
-public class AdminController {
-
-    private final ReservationService reservationService;
-
-    public AdminController(ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
-
-    @PostMapping("/reservations")
-    private ResponseEntity<ReservationResponse> addAdminReservation(@RequestBody AdminReservationRequest request) {
-        ReservationResponse reservationResponse = reservationService.addAdminReservation(request);
-
-        return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
-                .body(reservationResponse);
-    }
-}
