+++ src/main/java/roomescape/controller/api/ThemeController.java
@@ -1,23 +1,17 @@
 package roomescape.controller.api;
 
-import java.net.URI;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import roomescape.dto.request.ThemeRequest;
 import roomescape.dto.response.ThemeResponse;
 import roomescape.service.ThemeService;
 
-@RestController
 @RequestMapping("/themes")
+@RestController
 public class ThemeController {
 
     private final ThemeService themeService;
@@ -32,23 +26,6 @@ public class ThemeController {
 
         return ResponseEntity.ok()
                 .body(responses);
-    }
-
-    @PostMapping
-    public ResponseEntity<ThemeResponse> addTheme(@RequestBody ThemeRequest request) {
-        ThemeResponse response = themeService.addTheme(request);
-        URI location = URI.create("/themes/" + response.id());
-
-        return ResponseEntity.created(location)
-                .body(response);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
-        themeService.deleteThemeById(id);
-
-        return ResponseEntity.noContent()
-                .build();
     }
 
     @GetMapping("/rankings")
