+++ src/main/java/roomescape/controller/api/admin/AdminThemeController.java
@@ -1,0 +1,43 @@
+package roomescape.controller.api.admin;
+
+import java.net.URI;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import roomescape.dto.request.ThemeRequest;
+import roomescape.dto.response.ThemeResponse;
+import roomescape.service.ThemeService;
+
+@RequestMapping("/admin/themes")
+@RestController
+public class AdminThemeController {
+
+    private final ThemeService themeService;
+
+    public AdminThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<ThemeResponse> addTheme(@RequestBody ThemeRequest request) {
+        ThemeResponse response = themeService.addTheme(request);
+        URI location = URI.create("/themes/" + response.id());
+
+        return ResponseEntity.created(location)
+                .body(response);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
+        themeService.deleteThemeById(id);
+
+        return ResponseEntity.noContent()
+                .build();
+    }
+}
