+++ src/main/java/roomescape/controller/api/admin/AdminTimeController.java
@@ -1,0 +1,44 @@
+package roomescape.controller.api.admin;
+
+import java.net.URI;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import roomescape.dto.request.ReservationTimeRequest;
+import roomescape.dto.response.ReservationTimeResponse;
+import roomescape.service.ReservationTimeService;
+
+@RequestMapping("/admin/times")
+@RestController
+public class AdminTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public AdminTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationTimeResponse> addTime(@RequestBody ReservationTimeRequest request) {
+        ReservationTimeResponse response = reservationTimeService.addReservationTime(
+                request);
+        URI location = URI.create("/times/" + response.id());
+
+        return ResponseEntity.created(location)
+                .body(response);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTime(@PathVariable("id") Long id) {
+        reservationTimeService.deleteReservationTimeById(id);
+
+        return ResponseEntity.noContent()
+                .build();
+    }
+}
