+++ src/test/java/roomescape/acceptance/admin/ReservationFilterAcceptanceTest.java
@@ -96,7 +96,7 @@ class ReservationFilterAcceptanceTest extends BaseAcceptanceTest {
 
         return RestAssured.given().log().all()
                 .cookie("token", adminToken)
-                .when().get("/reservations/filter" + path)
+                .when().get("/admin/reservations/filter" + path)
                 .then().log().all()
                 .statusCode(HttpStatus.OK.value())
                 .extract().as(memberListFormat);
